Rewrite ActionsService.
