+++ Sources/TapestryKit/Commands/ActionsCommand.swift
@@ -1,29 +1,18 @@
-import TSCUtility
+import ArgumentParser
 import TSCBasic
 import TapestryCore
 import Foundation
 
 /// List all available actions
-final class ActionsCommand: NSObject, Command {
-    static var command: String = "actions"
-    static var overview: String = "Show available actions"
-    
-    required init(parser: ArgumentParser) {
-        parser.add(subparser: ActionsCommand.command, overview: ActionsCommand.overview)
+struct ActionsCommand: ParsableCommand {
+    static var configuration: CommandConfiguration {
+        CommandConfiguration(
+            commandName: "actions",
+            abstract: "Show available actions"
+        )
     }
     
-    func run(with arguments: ArgumentParser.Result) throws {
-        printActions()
-    }
-    
-    private func printActions() {
-        Action.allCases.forEach {
-            switch $0 {
-            case .docsUpdate:
-                Printer.shared.print("docs-update\tUpdate docs with a given version")
-            case .dependenciesCompatibility:
-                Printer.shared.print("compatibility\tCheck compatibility with given dependency managers")
-            }
-        }
+    func run() throws {
+        try ActionsService().run()
     }
 }
